dom/webgpu/tests/mochitest/test_buffer_unmap_on_invalid_device.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <script>
      ok(
        SpecialPowers.getBoolPref("dom.webgpu.enabled"),
        "Pref should be enabled."
      );

      async function testBody() {
        const adapter = await navigator.gpu.requestAdapter();
        const device = await adapter.requestDevice();

        const bufferRead = device.createBuffer({
          size: 4,
          usage: GPUBufferUsage.MAP_READ | GPUBufferUsage.COPY_DST,
        });
        const bufferWrite = device.createBuffer({
          size: 4,
          usage: GPUBufferUsage.COPY_SRC,
          mappedAtCreation: true,
        });

        const tryCatchOkMsg = "unmapping buffer with invalid device should not throw"
        try {
          device.addEventListener('uncapturederror', (event) => {
            ok(false, "device should not report any uncaptured errors")
          })
          device.pushErrorScope("internal");
          device.pushErrorScope("out-of-memory");
          device.pushErrorScope("validation");
          bufferRead.unmap();
          bufferWrite.unmap();
          ok(!await device.popErrorScope(), "device should not report any errors")
          ok(!await device.popErrorScope(), "device should not report any errors")
          ok(!await device.popErrorScope(), "device should not report any errors")
          ok(true, tryCatchOkMsg);
        } catch (e) {
          ok(false, tryCatchOkMsg);
        }
      }

      SimpleTest.waitForExplicitFinish();
      testBody()
        .catch(e => ok(false, "Unhandled exception " + e))
        .finally(() => SimpleTest.finish());
    </script>
  </body>
</html>
